The test take into account the changes in SNPDB/Obj.pm
get_SeqFeature_by_id now returns only SNPs which are mapped to a clone.
The returned Variation objects have start, end and strand are relative to the clone.
The clone accession_version is stored by ->seqname method.

Added 25 nucleotide flanking sequences to the returned Variation objects.
The methods are ->upStreamSeq and ->dnStreamSeq.