130 release updates. The BodyMap file has changed quite a lot... be careful about that in the future. See ensembl-doc/sage before ypu start mapping - it's quite complete descri[tion now. We seem to be mapping more and more novel genes which is a great thing.
Internal id/stable id is a bit ackward - I am using a script which is not commited because the change-over always takes plca at a different moment - depending on whether botj ENSU and ELUS mapping works on internal ids or just ENSU. But conversion is easy - just dump a table, make a hash and process a file with $stable_id{$internal_id} in place of mere $internal_id.

Good luck for 140 release when I waont be here.