Change to method get_ccds_id:
Previously would only get id if it did not exist. Now also gets the id if it exists but does not match /^CCDS/. This is needed because the new get_ccds_uid sets the ccds id as ccds_uid (eg. 1234) and not as the string pattern we expect (eg. CCDS1234.1). We need to overwrite the ccds_id to match /^CCDS/.